Add functionality for fallback on login
If the `STS_AWS_REGION` env var is present the STS endpoint in signed request will be updated to have a URI of `http://sts.<region>.amazonaws.com`.
For this to work correctly, the `auth/aws/config/client` path needs to be configured with the correct values for `sts_endpoint` and `sts_region`.

This change will allow the client to fall back to trying to login using the default, global endpoint of `https://sts.amazonaws.com` in situations where the `STS_AWS_REGION` has been configured but vault isn't configured.